requirements.txt
PySide6~=6.8.1.1
python-dotenv~=1.0.1
supabase~=2.11.0
bcrypt~=4.2.1